meta-analysis/backend/requirements.txt
Flask==3.1.0
Flask-Cors==3.0.10
requests==2.32.3
gunicorn==23.0.0
google-generativeai
reportlab